+++ .github/workflows/python.yml
@@ -36,7 +36,7 @@ jobs:
           args: --release --out dist --find-interpreter
           sccache: 'true'
           manylinux: auto
-          working-directory: /home/runner/work/anise/anise-py 
+          working-directory: anise-py 
       - name: Upload wheels
         uses: actions/upload-artifact@v3
         with:
@@ -60,7 +60,7 @@ jobs:
           target: ${{ matrix.target }}
           args: --release --out dist --find-interpreter
           sccache: 'true'
-          working-directory: /home/runner/work/anise/anise-py 
+          working-directory: anise-py 
       - name: Upload wheels
         uses: actions/upload-artifact@v3
         with:
@@ -100,7 +100,7 @@ jobs:
         with:
           command: sdist
           args: --out dist
-          working-directory: /home/runner/work/anise/anise-py 
+          working-directory: anise-py 
       - name: Upload sdist
         uses: actions/upload-artifact@v3
         with:
