+++ anise/tests/almanac/mod.rs
@@ -14,15 +14,9 @@ fn test_load_ctx() {
     let dataset = convert_tpc("../data/pck00008.tpc", "../data/gm_de431.tpc").unwrap();
 
     // Load BSP and BPC
-    let ctx = Almanac::default();
+    let ctx = Almanac::new("../data/de440.bsp").unwrap();
 
-    let spk = SPK::load("../data/de440.bsp").unwrap();
-
-    let mut loaded_ctx = ctx
-        .with_spk(spk)
-        .unwrap()
-        .load("../data/earth_latest_high_prec.bpc")
-        .unwrap();
+    let mut loaded_ctx = ctx.load("../data/earth_latest_high_prec.bpc").unwrap();
 
     loaded_ctx.planetary_data = dataset;
